Finishing touches + CSS

+++ projeto/action_edit_event.php
@@ -2,11 +2,20 @@
 include_once('database/events.php');
 include_once("database/upload.php");
 
-$image_path = uploadImage($_FILES['image'], 'resources/images/uploaded/');
-if ($image_path != 'error') {
-    echo '<p> Final path: ' . $image_path;
-    editEvent($_POST['id'], $_POST['date'], $_POST['desc'],
-              $_POST['type'], $image_path, $_POST['private_flag']);
+if (isset($_POST['confirm_btn'])) {
+    $priv = 0;
+    if ($_POST['privacy'] == 'Private') $priv = 1;
+
+    $image_path = uploadImage($_FILES['image'], 'resources/images/uploaded/');
+    if ($image_path == 'no_file') {
+        editEventNoImage($_POST['id'], $_POST['date'], $_POST['desc'],
+                         $_POST['type'], $priv);
+    }
+    else if ($image_path != 'error') {
+        echo '<p> Final path: ' . $image_path;
+        editEvent($_POST['id'], $_POST['date'], $_POST['desc'],
+                  $_POST['type'], $image_path, $priv);
+    }
 }
 
 header('Location: ' . $_COOKIE['redirect']);
